linux-gen: deduplicate arch-specific files
Rewrite ARCH-specific files so that there is no significant code
duplication between different arch files and default files set.
