+++ platform/linux-generic/Makefile.am
@@ -68,22 +68,6 @@ odpapiinclude_HEADERS = \
 		  $(srcdir)/include/odp/api/traffic_mngr.h \
 		  $(srcdir)/include/odp/api/version.h
 
-if ARCH_IS_ARM
-odpapiinclude_HEADERS += $(srcdir)/arch/arm/odp/api/cpu_arch.h
-endif
-if ARCH_IS_AARCH64
-odpapiinclude_HEADERS += $(srcdir)/arch/aarch64/odp/api/cpu_arch.h
-endif
-if ARCH_IS_MIPS64
-odpapiinclude_HEADERS += $(srcdir)/arch/mips64/odp/api/cpu_arch.h
-endif
-if ARCH_IS_POWERPC
-odpapiinclude_HEADERS += $(srcdir)/arch/powerpc/odp/api/cpu_arch.h
-endif
-if ARCH_IS_X86
-odpapiinclude_HEADERS += $(srcdir)/arch/x86/odp/api/cpu_arch.h
-endif
-
 odpapiplatincludedir= $(includedir)/odp/api/plat
 odpapiplatinclude_HEADERS = \
 		  $(builddir)/include/odp/api/plat/static_inline.h \
@@ -238,26 +222,46 @@ __LIB__libodp_linux_la_SOURCES = \
 			   odp_weak.c
 
 if ARCH_IS_ARM
-__LIB__libodp_linux_la_SOURCES += arch/arm/odp_cpu_arch.c \
-				  arch/arm/odp_sysinfo_parse.c
+__LIB__libodp_linux_la_SOURCES += arch/default/odp_cpu_arch.c \
+				  arch/default/odp_cpu_cycles.c \
+				  arch/default/odp_global_time.c \
+				  arch/default/odp_sysinfo_parse.c
+arch_odp_headers = $(srcdir)/arch/arm/odp/api/cpu_arch.h
 endif
 if ARCH_IS_AARCH64
-__LIB__libodp_linux_la_SOURCES += arch/aarch64/odp_cpu_arch.c \
-				  arch/aarch64/odp_sysinfo_parse.c
+__LIB__libodp_linux_la_SOURCES += arch/default/odp_cpu_arch.c \
+				  arch/default/odp_cpu_cycles.c \
+				  arch/aarch64/odp_global_time.c \
+				  arch/default/odp_sysinfo_parse.c
+arch_odp_headers = $(srcdir)/arch/aarch64/odp/api/cpu_arch.h
 endif
 if ARCH_IS_MIPS64
 __LIB__libodp_linux_la_SOURCES += arch/mips64/odp_cpu_arch.c \
+				  arch/default/odp_cpu_cycles.c \
+				  arch/default/odp_global_time.c \
 				  arch/mips64/odp_sysinfo_parse.c
+arch_odp_headers = $(srcdir)/arch/mips64/odp/api/cpu_arch.h
 endif
 if ARCH_IS_POWERPC
-__LIB__libodp_linux_la_SOURCES += arch/powerpc/odp_cpu_arch.c \
+__LIB__libodp_linux_la_SOURCES += arch/default/odp_cpu_arch.c \
+				  arch/default/odp_cpu_cycles.c \
+				  arch/default/odp_global_time.c \
 				  arch/powerpc/odp_sysinfo_parse.c
+arch_odp_headers = $(srcdir)/arch/powerpc/odp/api/cpu_arch.h
 endif
 if ARCH_IS_X86
 __LIB__libodp_linux_la_SOURCES += arch/x86/cpu_flags.c \
 				  arch/x86/odp_cpu_arch.c \
+				  arch/default/odp_cpu_cycles.c \
+				  arch/x86/odp_global_time.c \
 				  arch/x86/odp_sysinfo_parse.c
+arch_odp_headers = $(srcdir)/arch/x86/odp/api/cpu_arch.h
+noinst_HEADERS += $(srcdir)/arch/x86/cpu_flags.h
 endif
+
+noinst_HEADERS += $(srcdir)/arch/default/odp/api/cpu_arch.h
+
+odpapiinclude_HEADERS += $(arch_odp_headers)
 
 if HAVE_PCAP
 __LIB__libodp_linux_la_SOURCES += pktio/pcap.c
