+++ platform/linux-generic/arch/default/odp_cpu_arch.c
@@ -12,10 +12,7 @@
 #include <time.h>
 
 #include <odp/api/cpu.h>
-#include <odp/api/hints.h>
-#include <odp/api/system_info.h>
 #include <odp_debug_internal.h>
-#include <odp_time_internal.h>
 
 #define GIGA 1000000000
 
@@ -38,29 +35,4 @@ uint64_t odp_cpu_cycles(void)
 	cycles += (ns * hz) / GIGA;
 
 	return cycles;
-}
-
-uint64_t odp_cpu_cycles_max(void)
-{
-	return UINT64_MAX;
-}
-
-uint64_t odp_cpu_cycles_resolution(void)
-{
-	return 1;
-}
-
-int cpu_has_global_time(void)
-{
-	return 0;
-}
-
-uint64_t cpu_global_time(void)
-{
-	return 0;
-}
-
-uint64_t cpu_global_time_freq(void)
-{
-	return 0;
 }
