+++ platform/linux-generic/arch/powerpc/odp_cpu_arch.c
@@ -39,28 +39,3 @@ uint64_t odp_cpu_cycles(void)
 
 	return cycles;
 }
-
-uint64_t odp_cpu_cycles_max(void)
-{
-	return UINT64_MAX;
-}
-
-uint64_t odp_cpu_cycles_resolution(void)
-{
-	return 1;
-}
-
-int cpu_has_global_time(void)
-{
-	return 0;
-}
-
-uint64_t cpu_global_time(void)
-{
-	return 0;
-}
-
-uint64_t cpu_global_time_freq(void)
-{
-	return 0;
-}
