+++ platform/linux-generic/arch/powerpc/odp_global_time.c
@@ -1,0 +1,15 @@
+
+int cpu_has_global_time(void)
+{
+	return 0;
+}
+
+uint64_t cpu_global_time(void)
+{
+	return 0;
+}
+
+uint64_t cpu_global_time_freq(void)
+{
+	return 0;
+}
